added features from 1 to 4 and improved code readability and efficiency
